pyproject.toml
[tool.poetry]
name = "socar-data-car-image-uploader"
version = "0.1.0"
description = ""
authors = ["nuj <nuj@socar.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
reflex = "^0.6.3"
google-cloud-bigquery = "^3.26.0"
python-dotenv = "^1.0.1"
google-cloud-storage = "^2.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
